pyproject.toml
[tool.poetry]
name = "dw-zero"
version = "0.1.0"
description = ""
authors = ["levyvix <levy.vix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
yfinance = "^0.2.40"
black = "^24.4.2"
dbt-postgres = "^1.8.2"
ipython = "^8.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
